Fix scrubber start & not scrubbing when not daemon
1) Scrubber not starting
When the scrubber is run from the command line or via cron, it would try
to call it's app object which resulted in a TypeError.

2) Scrubber not scrubbing
In the same usage scenario as above, the delete jobs added to the
starmap are never started.  A slight delay is enough to start these.

Also add a functional test designed to excercise the scrubber when run
in "regular mode", which covers both these cases.  I've run the test
successfully with 1000 images, which should be enough to prove the
sleep(0.1) does the trick.

Fixes bug 1052537.

Change-Id: If29704d8a425194085b59b7dfd265ffaef16302c