Enable multi-processing by default.
In realistic production environments, multi-processing is likely
to be enabled.

A default setting of one worker would allow the multi-processing
codepath to be followed by default, without changing the number
of processes actually servicing API requests (i.e. the single
parent process will just wait on the child).

For most functional tests, we continue to explicitly configure
zero workers, so as to avoid spawning twice as many glance-api
processes during a test run.

Change-Id: I1c327805f8fb975a3175e9a8b40aeddd21d5ef91