Fix scrubber not scrubbing with swift backend
It seems the fix for bug #1052537 was incomplete, as I have found
glance-scrubber is still not completing it's work when run against a
swift backend.

With more investigation it seems the proper way to execute a
starmap/imap is to consume it within a loop. This can be replicated by
taking the simple web crawler example from http://eventlet.net/doc/ and
removing the for loop from around the imap call.

Also:
* Add new test to check this failure against Swift (respects the
  GLANCE_TEST_SWIFT_CONF environment variable).

Fixes bug 1056127

Change-Id: I2d289666b99f3216a6d424dbbacbd2d10899263d