Add insecure option to registry https client
This fixes LP bug #1063793.

This adds a new API config option registry_client_insecure
(default false) which optionally allows not to specify a ca
file for registry connections from the API server.
The default is false so existing behaviour is
unchanged unless this option is specified.
Adding this option makes the registry symmetric wrt the API
with which you can connect using clients with --insecure

Change-Id: I4fa462cd68afaf14f7d7de6c572980a41c0849d6