Return 403 on images you can see but can't modify
Visible images return 404 (Not Found) when you try to modify them and
are not allowed. This patch changes this return to 403 Forbidden which
more accurately reflects the situation, in light of the fact that we are
not trying to hide the existence of the image in this case.

Fixes bug 1078520

Change-Id: I70e6e273aeaef51dad40cf001308d3a817bdced5