Add an image domain model and related helpers.
This patch is the first in a series refactoring the logic around image
manipulation. In this patch, we add a domain module that has Images and
ways of constructing them. In anticipation of proxy classes added in
subsequent patches, we also define some base classes to make proxying
simpler.

Partially implements bp:glance-domain-logic-layer

Change-Id: Ie72f2f86cfcbd15e41c32deed3626290f2bf0e6b