Stop logging all registry client exceptions
Not all exceptions need to be logged as ERROR in the glance-api
log when - for instance, a user asking for an image that does not
exist doesn't mean an ERROR actually occurred. This logs the exception
name at INFO level. An ERROR won't be logged at all, even in
the case of a 500-level response from glance-registry. An operator
would be alerted with an ERROR log in the glance-registry log in that
case.

Fixes bug 1079211.

Change-Id: Ie2bba3105f8d9f3de97bc50132fb046baf710a90