Adding a means for a glance worker to connect back to a pydevd debugger.
That patch allows a developer to remotely run a pydev debugger and a
glance worker process connect back to it.  Two configuration options
have been added:
    pydev_worker_debug_host <host>. Default is None
    pydev_worker_debug_port <port>. Default is 5678
The behavior is enabled if pydev_worker_debug_host is not None.  If
a pydev connection fails to be esstablished an exception is raised.

This patch will allow remote debugging as well as more seemless
debugging environments with IDEs like pycharm.

Change-Id: I7d3f60e373632b256f0e914f18204ce223a3486e