Rearrange db tests in prep for testr.
The way discover does test discovery will find any classes that subclass
the unittest base class. This means that db.base was getting picked up as
a set of tests that need to be run, which is bogus, as it's intended to just
be a base class.

Split it up a little bit differently to make it work as intended when
python test discovery is run outside of how nose things of things.

Part of blueprint grizzly-testtools

Change-Id: Iede92452d0c48224a1f80288c2bf2693224dc56f