Stub out find_file... fix policy.json test issue.
Updates test_policy.py so that we stub out find_file
which is used to find the *default* policy file. This
fixes an issue in the test_policy_file_default_not_found
test where it *may* fail when executed on a machine with
Glance package(s) installed (thus created a file in /etc/glance/
that isn't readable by the test user).

Fixes LP Bug #1098687.

Change-Id: If2912d40608ffb07f457454fdc6ebc56c62f6eaf