Use more-specific value for X-Object-Manifest header.
When uploading a large image to Swift, the segments are named
glance/<imageid>-<segmentnumber>. However, the value of
X-Object-Manifest is only glance/<imageid> without the trailing
dash. This can lead to odd behavior when there are other objects that
start with <imageid>, but are not actually segments of the image.

For example, in bug 1112653, a user was using python-swiftclient to
delete an old, segmented image with id 10. However, they also had an
image with id 100b2316-9187-48d8-b327-dded22d1dc71-00001, and since
"10" is a prefix of "100b2316-9187-48d8-b327-dded22d1dc71-00001",
swiftclient thought that the second image was just another segment of
image 10 and deleted it.

Further, had someone issued a GET request for /glance/10, the
resulting response would have contained the bytes from 10's segments
*plus* the bytes from 100b2316-9187-48d8-b327-dded22d1dc71-00001,
which is definitely not the right behavior.

Obviously, this change doesn't go back and fix up old
manifests. However, it should help prevent this problem from
recurring.

Change-Id: I37577fde7751ad76ce543bc59eb7ee8de2187ac3