Fix issues with migration 012
This commit fixes 3 issues with migration 012:

First for the sqlite downgrade case the schema was being updated
before the data, which made it impossible to be applied. This
reverses the order of those calls correcting the issue.

Second, when the generate_uuid method was moved to one synced
from oslo the import was never updated in migration 012 which
was using generate_uuid(). This commit updates the imports and
uses of generate_uuid() to use the method in oslo.

Thirdly, this patch fixed the downgrade path to work for an image
table with more than one row.

Change-Id: Ib0d3d67e924748863dfdf8eb9e3257ec108ecf68