Add GridFS store
This patch adds support for GridFS to glance. GridFS is 10gen's, mongodb
based, distributed filesystem.

The store requires 2 options in order to work correctly:
    * mongodb_store_uri
    * mongodb_store_db

Note: mongodb was used in the parameters name's even though this store
supports `gridfs://` schemas (not official) in order to keep consistency
with mongodb's documentation and connection parameters.

GridFS StoreLocations:

This store implements gridfs locations with the form `gridfs://<IMAGE>`.
Note that there's no connection information in those uris, which was
consciously omitted for security reasons. Perhaps, it could be possible
to make that information available by setting a configuration parameter
that specifies so.

Implements blueprint gridfs-store

Change-Id: I38f361ac8421f0445dd37800c840441a46cf9fdf