Filter images by status and add visibility shared
This patch allows the user to filter images by status.
Whenever a user does an image list, the following are displayed:
(1) public images
(2) images owned by the user
(3) images whom the user is a member of and has set status to accepted
The user can change the status in (3) and also choose to display
the images the user has been requested to be a member of,
rejected to be a member of, or all images irrespective of the status
set by user

If visibility filter can be shared, public or private. Where shared
filter displays images which the user is a member of. This can be
used in conjunction with member_status

Related to bp glance-api-v2-image-sharing

Change-Id: I052c0947f9ad17ef1b14af0ad0f51b886afd0d7a