Swallow UserWarning from glance-cache-manage
Fixes bug 1129445

Previously, we exposed internal technical debt as an apparent
user-facing warning.

Now the warning is swallowed in the internal case, but continues
to be seen by external users of this library.

Note that this fix is not relevant on master, as that the warning
no longer occurs in that case.

Change-Id: I30a835e0024acac8841b4746968abe7cae9e14cb