Fix visibility on db image_member_find
Before this change, image_member_find was not paying any attention to
whether or not the requesting context was authorized to view the image
memberships. With this change, image membership results are always
limited to either images owned by the context tenant, or images shared
with the context tenant.

Fixes bug 1114821

Change-Id: I9cabe218429e32c7e2db932ced141e056e10d1c5