Fix rabbit_max_retry
First, the configvar is defined as a StrOpt. This means that the check
against the integer retry_attempts will fail.

Secondly, rabbit_max_retries is tested as a boolean before comparing
against retry_attempts. With the value 0, this would be false and the
setting rabbit_max_retries=0 would actually retry indefinitely, which is
not the obvious behavior. This patch will make the default not to
retry at all (which was the implied, but not actual, behavior)

Fixes bug 1152866

Change-Id: Ic39d708f0294fc5d919cb2a5d9fae532128ef17e