Remove internal store references from migration 015
Migration 015 was accessing a object internal to the swift store driver
and changing its behavior in order to do the migration.  This worked but
it made the assumption that the driver logic would not change without
also changing the migration code.  These two should not be dependent
upon each other.  This patch removes all dependencies on swift from the
migration code.

Fixes bug: 1025746
Change-Id: Ic9559ae809b8784d0992cb5275d3911a725110c8