Ensure repeated member deletion fails with 404
Fixes bug 1157427

If an image membership is deleted more than once, the second and
subsequent deletions would be expected to return a 404 Not Found
response code, as opposed to 204 No Content.

Previously, the can_show_deleted logic got in the way, such that
for an admin caller the previously deleted membership was considered
still extant.

We now override the show_deleted logic for deletion, as it makes no
sense to consider the existence of previously deleted database rows
when another deletion request is issued.

Note that there is no corresponding proposal to master as this logic
has been refactored and this no longer affects master.

Change-Id: If8200bbb74e46c814c9f8fa6b4c636e7b022ab6c