Check if creds are present and not None
Currently, KeystoneStrategy checks whether the required credentials are
present in the creds dict but it doesn't check whether they are valid or
not.

This patch checks whether the required creds are present and not None
otherwise a MissingCredentials exception will be raised.

Note: No need for checking parameters' values types since they'll be
instances of basestring once set otherwise they'll be None.

Fixes bug 1157765

Change-Id: I664a604c3cbf2fca60a88c4d887cd9a4b678c8a5