Add placeholder migrations to allow backports
Inspired by the nova blueprint backportable-db-migrations,
we reserve a range of migration slots to allow for idempotent
and standalone migrations to be backported to stable/grizzly
during the Havana release cycle.

Nova have reserved 10 slots and the historical ratio of
nova:glance migrations is circa 8:1. Hence the thought that
a total of 3 placeholder slots should more than suffice for
the Havana release cycle.

Change-Id: I576be7520f960e4cb4dcc6928035595fab7e459f