Remove internal store references from migration 017
Migration 017 was accessing a object internal to the swift store driver
and changing its behavior in order to do the migration. This worked but
it made the assumption that the driver logic would not change without
also changing the migration code.  This patch removes all dependencies on
the swift driver from the migration code.  This patch is smilar to
https://review.openstack.org/#/c/24552/.

Fixes bug: 1155749
Change-Id: Iffd800f62d61bf7608f7be5b23d76d702e6d4064