Fix functional test 'test_copy_from_swift'
This test fails if a real swift system is specified in a config file
and passed to the tests via the GLANCE_TEST_SWIFT_CONF environment
variable. The fix is to quote the creds in the uri and also to
fix get_swift_uri() to handle uris of the form '@https://'.

Fixes bug 1158404

Change-Id: Ic091f77a83696e6413aea047e805ac11487aba39