Make is_public an argument rather than a filter.
There is disagreement between the v1 api and v2 api about what an
'is_public' filter ought to mean. Version 1 uses 'is_public'
to filter whether or not image.is_public == True or False. Version 2
uses 'visibility' for this purpose. But in addition, it is theoretically
possible that 'is_public' is a valid property filter in v2.

With this change, v1 can pass is_public in as an argument to
image_get_all. V2 will continue to pass is_public through filters.

fixes bug: 1157988
fixes bug: 1154760

Change-Id: I716cb8b4f207f8526125ed360f5a24a6630c1e67