Directly verifying that time and socket are monkey patched
This tests makes sure that thew wsgi object properly monkey patched
two needed modules, socket and time.  In bug 988099 it was noted that
the time module was not being monkey patched.  A test was added that
unfortunately introduced a bug and a patch to revert it is here:
https://review.openstack.org/#/c/25107/ in order to fix bug 1158247.
This patch allows for test_server_monkey_patch.py to be removed
without losing the verification of that functionality.

Fixes bug 1158247
Change-Id: I8da6b04d0cc0dae00ba52e5b7350a5e01a76cadb