glance-manage should not require glance-registry.conf
In v2 the glance-registry service is no longer needed.  Thus
it is possible that a setup would not include a configuration
file for it at all.  Further it is possible that glance-api and
glance-registry be run on different machines such that neither
configuration file is available to the other.  Therefore glance-manage
should be able to pull the information out of either configuration
file provided that the overriding command line options:
--sql_connection or --config-file are not used.  This patch makes
it so glance-manage will first look in the glance-registry.conf file
for db information (as this is the legacy behavior) and then in
the glance-api.conf file if that file does not exist.

Fixes bug: 1046576
Change-Id: Ida68a5c03eb3d188b2b997296cd45730b968ef12