Functional tests display the logs of the services they started
Function tests fork out processes that run glance-{api,registry,etc}.
These services are started with a conf file that directs logging
to a different file.  Unfortunately that logged information is not
recorded by the test frameworks that run these tests.  This change
opens the log file and dumps it out to a logger created for that
service.  In this way a developer can inspect the logs of all the
services created by a failed tests using the test frameworks interface
(like Jenkins web interface).

Fixes bug: 1162933

Change-Id: Ie699baaa376ca4ec848a8600f9fc5da6484b24f0