Invalid reference to self in functional test test_scrubber.py
In the test_scrubber.py the function wait_for_scrub() makes a
reference to self even tho it is not a member of any class and
has no such reference in its scope.  This patch makes that
function a method in the test class so that it can assert the
failure which it intends.

fixes bug: 1163617

Change-Id: I453c8097363dc2f7577fd807ffe8cbc613078258