Call monkey_patch before other modules are loaded
This patch moves the monkey_patch call into the glance scripts that
needs libs to be monkey_patched. This seems to be a good practice and
has given good results for other projects as well.

This change is needed  to avoid loading / initializing modules
before other libs are patched. For example: logging module.

Fixes bug 1163722

Change-Id: Idee486851606d765fb11598967a4f7d56cf46c97