Call os.kill for each child instead of the process group
The patch sets the process group id to its process id. Note that this
change raises an OSError in Glance's tests because they are run under
Popen and using glance-control, which changes children's session ids and
other parameters. That for, setpgid is inside a try / except and falls
back to 0.

Fixes bug 1068051

Change-Id: Ie9ed044e08053d0067396138de0bd124326aa1a5