Fix Qpid test cases
The Qpid notification test cases do not currently pass.
By default they are skipped because qpid is not part of
test-requires, so it isn't automatically installed.  This fix
updates the test cases to work correctly, but does not add
qpid to test-requires because I'm unsure what impacts that
might have on other platforms.

This is part of the fix for bug 1164681

Change-Id: I40757b5de99559b4ff090e0d249f8624c2e2be10