Create package for registry's client
Currently, Registry's client doesn't have a version and It's being
imported from glance.registry.(__init__|client).

As part of the blueprint registry-api-v2 the registry's client will be
modified. That for, it is necessary to keep backward compatibilities
with older versions until the db-registry-driver is completed.

This code might be removed / re-factored in the future once glance's api
v1 will be migrated to use the db-registry-driver.

Change-Id: I3f27cec5b6e5f857d7aa94731305629190ae642e