Remove glance-control from the test suite
This patch removes the use of glance-control from the test suite. Instead
of forking out glance-control and letting it fork out the other programs,
we will now fork out the other programs directly. The file which tested
glance-control has also been removed. While we will still have
glance-control as part of the distribution for a bit and thus it should
still be tested, that test was doing so in a way that depended upon the
forking behavior previously found in the test suite.

This patch also requires the patch
https://review.openstack.org/#/c/26076/ on which it is dependent. As a
side effect it tests for the problem found in bug: 1068051

Additionally, this fixes a race condition. Previously each functional
test called cleanup() in tearDown(). Cleanup would open a pid file and
send a SIGTERM to the processes (api, registry, scrubber). However,
it did not wait for this process to die and get cleaned up. The signal
handler in the processes is over ridden so the death is not instant.
This leaves a chance for server from a previous test to be actively
listening on a port when the next test starts.

fixes bug: 1173415
blueprint: refactoring-better-faster-stronger-functional-tests

Change-Id: Iabdb6fdc13b5f1993590b5a801a0e89df1a544c7