Redirects requests from /v# to /v#/ with correct Location header
This is a continuous fix to bug #1166957

VersionNegotiationFilter pops off the version string of PATH_INFO,
the popped segment is pushed onto SCRIPT_NAME, which results in
dual version strings (e.g. '/v1/v1/') in the 'Location' header
of a redirect response. This may result in a further 404 error
if the http client will automatically follow the returned
redirect.

Fix this problem by adding a private method which does not
push the popped segment onto SCRIPT_NAME.

Change-Id: Ie0c244af38534dde2058828fbff840a845cc7c90
Fixes: bug #1175815