rbd: remove extra str() conversions and test with unicode
librbd doesn't handle unicode, so strings passed to it must be plain
ascii. Originally the rbd store did this before each call, but since
3aea9320d20d0291c5b4b027a585370d8e658a95 it only needed to convert to
ascii in StoreLocation and add(). Remove the extraneous calls to
str(), and add a test that uses a unicode image_id and uri to make
sure unicode isn't passed to librbd anywhere.

Change-Id: Ib7656126e044775f516fa6665e28f6d474820665
Signed-off-by: Josh Durgin <josh.durgin@inktank.com>