Port slow, overly assertive v1 functional tests to integration tests
This patch ports existing v1 api functional tests to a newly added group
of integration tests. The new tests take about 1/8th of the time of the
previous tests, thanks to not forking off api and registry server
processes.

One original functional test is preserved, to ensure that we are still
testing the basic v1 functional wireup.

Related to bp refactoring-better-faster-stronger-functional-tests

Change-Id: Ifcc76b4bf4e484a52558538e5398cad882e48e4c