Move sql_connection option into sqlalchemy package.
Since we're in the process of adding new backends that don't use
sql_connection, it makes sense to move that option into the sqlalchemy
package. That's what this patch does.

As for glance-manage, it now calls the add_cli_options directly from
sqlalchemy's api module.

This is part of the blueprint registry-db-driver

Change-Id: I890a5f707fe18f397cbeaadbac1d693ead2fdc14