Add location specific information to image locations db
This patch adds a column to the image_locations table.  The column
will store location specific information.  As a location is added
to an image there may be some meta data about that information that
needs to be stored with it.  For example, if it is a file:// URL,
information about the NFS server (host name, mount point) would be
needed in order for the URL to be useful.

blueprint: direct-url-meta-data
blueprint: multiple-image-locations

Change-Id: I6a96163abc8e4388e84cae2a8952f89d433210ff