Pass configure_via_auth down to auth plugin
Pass down the configure_via_auth value to the auth plugin to say whether or not
it should look in the service catalog for the glance registry endpoint. This
stops authentication by glance registry client from needlessly looking for an
"image" service endpoint when the user may not have it in their service catalog.

Fixes bug 1187888

Change-Id: I1c4a726ff0d6a345a27446500135f4f27cea5e39