Add custom RPC(Des|S)erializer to common/rpc.py
We need a custom way to serialize / de-serialize some types in the rpc
API. Some of this custom serializations are worth moving to Oslo and
other aren't. Before going down that road, it is worth testing them and
make them a bit more stable.

This patch introduces a serializer and a de-serializer for Glance's RPC
API.

Implements blueprint registry-db-driver

Change-Id: I70c429c65d67cbd7ba09176d5f8e4e8f27270c5b