Make db properties functions consistent with the DB API
Currently property related functions are not consistent with the
database API. As for the create operation, the API returns an
ImageProperty model instead of a dictionary and in the case of the
delete operation, the function expects a property instance instead of a
reference name.

This patch makes both calls consistent by returning a dictionary when a
property is created and receiving a property name instead of its
instance when a property needs to be deleted.

Fixes bug: #1199364
Implements blueprint registry-db-driver

Change-Id: Ibc7067cf712942610c7507f22f8895c2d531172a