DB Driver for the Registry Service
This patch adds a database driver that talks to a remote registry
service, which talks directly to a database back-end. The registry
service implements all the database API public functions that are
actually used from outside the API.

In order to test this patch correctly, it was necessary to add an
auth_token to each context used in the base tests and update that token
were needed.

NOTE: This patch adds tests for the above said driver but, it doesn't
add Visibility tests. The reason is that those tests are meant for
actual database back-ends. Given the "proxy" nature of this driver and
the fact that it doesn't do anything on the data returned by the remote
registry service, it isn't worth adding those tests. However, this
review does have support for DriverTests and implements each one of
them.

Implements blueprint registry-db-driver

Change-Id: I60f8e25c406c8e62b343267ba00f23299b19d4e5