File system store can send metadata back with the location.
This patch allows the glance operator to associate metadata with
the file system store.  The metadata will be returned to v2 users
with each location that was created by the file system store.  This
information can give clients context into how to consume the data.

The new configuration option: file system_store_metadata_file is
added.  This is a path to a JSON file.  The JSON document can
contain lists, dicts and unicode values only.

blueprint: direct-url-meta-data
blueprint: multiple-image-locations
docImpact
Change-Id: I3f040374455feb876e970a1813099598301774b5