Handle None value properties in glance-replicator
bug 1208417

In glance-replicator, all image properties in the master glance are
synced to the slave. But there are circumstances(kernel_id or
ramdisk_id, etc.) when image properties in the master are None.
glance-relicator doesn't handle this issue, thus all such properties
are trasnlated to the string 'None' in the slave glance, which is
incorrect.

This patch fixes this problem by add a check before post metadata to
the slave glance. All None properties are translated to '' before
posted to slave glance.

Change-Id: I9b674db907f707ac4dce28c8a7e3d81fcce61193