Use latest Oslo's version
The patch replaces current Oslo version with 1.2. The change is needed
to pull in a fix released in that specific version, which fixes the
issue #1175570. The fix we're interested in is giving priority to CLI
arguments instead of config files. Without this, commands like the one
below won't work.

    $ glance-manage --sql_connection=.... db_sync

Oslo's release notes:
http://docs.openstack.org/developer/oslo.config/#a3

Closes-Bug: #1175570
Change-Id: I4ea749cfaae0e18e035abe7b5d7e345dda7d476d